Added words and locals

+++ src/onda_dict.h
@@ -7,7 +7,7 @@
 typedef struct {
   char* key;
   size_t key_len;
-  uint32_t value;
+  uint64_t value;
 } onda_dict_slot_t;
 
 typedef struct {
@@ -21,10 +21,10 @@ void onda_dict_free(onda_dict_t* d);
 int onda_dict_get(onda_dict_t* d,
                   const char* key,
                   size_t key_len,
-                  uint32_t* out);
+                  uint64_t* out);
 void onda_dict_put(onda_dict_t* d,
                    const char* key,
                    size_t key_len,
-                   uint32_t value);
+                   uint64_t value);
 
 #endif // ONDA_DICT_H
